Switch to method-name convention for determining read vs write
Instead of explicitly classifying all Cosmos DB Container methods
as representing a Read or a Write, this classification is now done
just by looking at whether the method name starts with "Read" or not.
All of the relevant methods on Container follow a naming convention
of "<Verb><Noun>", so that enables this library to be more tolerant
of new Cosmos DB SDK versions that introduce new methods on Container
that did not exist at the time that this library was last published.